requirements.txt
accelerate==0.31.0
datasets==2.20.0
einops==0.8.0
filelock==3.13.1
flash-attn==2.5.9.post1
friendlywords==1.1.3
fsspec>=2023.1.0,<=2024.5.0
huggingface-hub>=0.23.4
jinja2>=3.1.3
Levenshtein==0.25.1
matplotlib>=3.10.0
numpy>=2.0.0
opencv-python>=4.12.0
packaging>=24.1
pandas==2.2.2
peft==0.10.0
pillow>=10.2.0
pyarrow==16.1.0
PyYAML>=6.0.1
rapidfuzz==3.9.3
requests>=2.32.3
ruff==0.4.9
safetensors==0.4.3
scikit-learn>=1.6.0
scipy>=1.15.0
timm==1.0.7
tokenizers>=0.19.1
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
transformers>=4.46.3
triton==2.3.1
trl>=0.15.0
typing-extensions>=4.12.2
wandb>=0.26.0
xxhash==3.4.1

# Optional for stage-3 ZeRO training. Install a version compatible with your CUDA/PyTorch stack.
deepspeed>=0.14.0
